Cleanup and restructure utils
