+++ lib/streamlit/delta_generator.py
@@ -32,6 +32,7 @@ from typing import (
 
 from typing_extensions import TypeAlias
 
+from lib.streamlit.elements.lib.form_utils import FormData, current_form_id
 from streamlit import (
     cli_util,
     config,
@@ -55,7 +56,6 @@ from streamlit.elements.doc_string import HelpMixin
 from streamlit.elements.empty import EmptyMixin
 from streamlit.elements.exception import ExceptionMixin
 from streamlit.elements.form import FormMixin
-from streamlit.elements.form_utils import FormData, current_form_id
 from streamlit.elements.graphviz_chart import GraphvizMixin
 from streamlit.elements.heading import HeadingMixin
 from streamlit.elements.html import HtmlMixin
