+++ lib/streamlit/elements/arrow.py
@@ -30,8 +30,8 @@ from typing import (
 
 from typing_extensions import TypeAlias
 
+from lib.streamlit.elements.lib.form_utils import current_form_id
 from streamlit import dataframe_util
-from streamlit.elements.form_utils import current_form_id
 from streamlit.elements.lib.column_config_utils import (
     INDEX_IDENTIFIER,
     ColumnConfigMappingInput,
