+++ lib/streamlit/elements/widgets/checkbox.py
@@ -18,7 +18,7 @@ from dataclasses import dataclass
 from textwrap import dedent
 from typing import TYPE_CHECKING, cast
 
-from streamlit.elements.form_utils import current_form_id
+from lib.streamlit.elements.lib.form_utils import current_form_id
 from streamlit.elements.lib.policies import (
     check_widget_policies,
     maybe_raise_label_warnings,
