+++ lib/streamlit/elements/widgets/file_uploader.py
@@ -20,8 +20,8 @@ from typing import TYPE_CHECKING, List, Literal, Sequence, Union, cast, overload
 
 from typing_extensions import TypeAlias
 
+from lib.streamlit.elements.lib.form_utils import current_form_id
 from streamlit import config
-from streamlit.elements.form_utils import current_form_id
 from streamlit.elements.lib.policies import (
     check_widget_policies,
     maybe_raise_label_warnings,
