+++ lib/streamlit/elements/widgets/multiselect.py
@@ -18,8 +18,11 @@ from dataclasses import dataclass, field
 from textwrap import dedent
 from typing import TYPE_CHECKING, Any, Callable, Generic, Sequence, cast
 
+from lib.streamlit.elements.lib.form_utils import current_form_id
+from lib.streamlit.elements.lib.utils import (
+    save_for_app_testing,
+)
 from streamlit.dataframe_util import OptionSequence
-from streamlit.elements.form_utils import current_form_id
 from streamlit.elements.lib.options_selector_utils import (
     check_and_convert_to_indices,
     convert_to_sequence_and_check_comparable,
@@ -44,9 +47,6 @@ from streamlit.proto.MultiSelect_pb2 import MultiSelect as MultiSelectProto
 from streamlit.runtime.metrics_util import gather_metrics
 from streamlit.runtime.scriptrunner import ScriptRunContext, get_script_run_ctx
 from streamlit.runtime.state import register_widget
-from streamlit.runtime.state.common import (
-    save_for_app_testing,
-)
 from streamlit.type_util import (
     T,
     is_iterable,
