+++ lib/streamlit/elements/widgets/select_slider.py
@@ -29,8 +29,9 @@ from typing import (
 
 from typing_extensions import TypeGuard
 
+from lib.streamlit.elements.lib.form_utils import current_form_id
+from lib.streamlit.elements.lib.utils import save_for_app_testing
 from streamlit.dataframe_util import OptionSequence, convert_anything_to_list
-from streamlit.elements.form_utils import current_form_id
 from streamlit.elements.lib.policies import (
     check_widget_policies,
     maybe_raise_label_warnings,
@@ -56,7 +57,6 @@ from streamlit.runtime.state import (
 )
 from streamlit.runtime.state.common import (
     RegisterWidgetResult,
-    save_for_app_testing,
 )
 from streamlit.type_util import T, check_python_comparable
 from streamlit.util import index_
