+++ lib/streamlit/elements/widgets/slider.py
@@ -33,7 +33,7 @@ from typing import (
 
 from typing_extensions import TypeAlias
 
-from streamlit.elements.form_utils import current_form_id
+from lib.streamlit.elements.lib.form_utils import current_form_id
 from streamlit.elements.lib.policies import (
     check_widget_policies,
     maybe_raise_label_warnings,
