+++ lib/streamlit/runtime/state/common.py
@@ -34,7 +34,7 @@ from typing import (
 
 from typing_extensions import TypeAlias, TypeGuard
 
-from streamlit import config, util
+from streamlit import util
 from streamlit.errors import (
     StreamlitAPIException,
 )
@@ -61,7 +61,6 @@ from streamlit.proto.TextInput_pb2 import TextInput
 from streamlit.proto.TimeInput_pb2 import TimeInput
 
 if TYPE_CHECKING:
-    from streamlit.runtime.scriptrunner_utils.script_run_context import ScriptRunContext
 
 
 # Protobuf types for all widgets.
@@ -244,9 +243,3 @@ def require_valid_user_key(key: str) -> None:
         )
 
 
-def save_for_app_testing(ctx: ScriptRunContext, k: str, v: Any):
-    if config.get_option("global.appTest"):
-        try:
-            ctx.session_state[TESTING_KEY][k] = v
-        except KeyError:
-            ctx.session_state[TESTING_KEY] = {k: v}
